refactor(app): 拆出全局弹层子组件（右键菜单/会话列表/飞行动画）
- 新增 components/overlays/overlayTypes.ts：AppOverlays props 与共享菜单状态类型（保留原导出路径）
- 新增 components/overlays/：EditFlyLayer、TabContextMenuOverlay、TerminalTabContextMenuOverlay、SessionListOverlay（各弹层渲染原样搬移）
- AppOverlays.tsx 553 -> 238 行
